lk-voice-agent/livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[elevenlabs,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]
